+++ src/channel.h
@@ -9,7 +9,7 @@ typedef struct {
 	root_node_t *subscribers;
 } channel_t;
 
-channel_t *create_channel(int id);
+channel_t *create_channel(uint64_t id);
 void destroy_channel(channel_t *channel);
 
 #endif
